Revamp keystroke detection system, again...
ncurses ncurses ncurses ncurses ncurses

So terminal input is cursed. I originally did not want to use the keypad
feature of ncurses because it has a lot of limitations, such as trying
to tell the difference between C-j and RET (and many more cases), the
way it handles ESC, etc. But in the end I gave in to using it as it made
most things easier.

See comments in controller.py for more details.