Private a bunch of attributes in the top level Ihmacs class.
These were attributes that were not and should not be mutated directly.
Added properties so they can be fetched the same way.

In addition, changed architecture from storing a pointer to the active
buffer to the index of the active buffer in the buffer list. This will
make it easier to change buffers down the line.

An active_buff property was added that returns the buffer object of the
active buffer. This allows for existing code to function exactly the
same.