Fix ncurses related crashes regarding input and output.
For input, limit the number of keys grabbed in a stroke to 2. I am using
keypad the only reason this would have extended otherwise was if someone
was pressing ESC (or one of the C- combos that produces 27 as well)
multiple times.

Also for input, fix the glaring holes in the unsightly if-elif tree, and
have an else statement to catch all other cases that don't fit.

For the output, just run `curses.update_line_cols()` in case the
terminal size has changed. This prevents the crashing.